# Fix performance with Merge table over huge number of MergeTree tables