--- a/src/Interpreters/ExpressionAnalyzer.h
+++ b/src/Interpreters/ExpressionAnalyzer.h
@@ -163,7 +163,7 @@ protected:
 
     const ASTSelectQuery * getSelectQuery() const;
 
-    bool isRemoteStorage() const;
+    bool isRemoteStorage() const { return syntax->is_remote_storage; }
 };
 
 class SelectQueryExpressionAnalyzer;