I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Fix performance of reading from `Merge` tables over huge number of `MergeTree` tables. Fixes #7748.

Detailed description / Documentation draft:
When `KeyCondition` is created, there is an invocation of `ExpressionAnalyzer::getConstActions`, which calls `IStorage::isRemote()`. But, when we read from `Merge` table, object of storage is `StorageMerge` in `query_info.syntax_analyzer_result` and `isRemote` iterates over all matched tables. So, there were O(n^2) iterations over tables, which is quite heavy operation.