@CurtizJ 2020-11-13
```
for i in {1..1000}; do clickhouse-client -q "create table if not exists mt_$i ENGINE = MergeTree ORDER BY key AS SELECT number AS key FROM numbers(1000)"; done

clickhouse-client -q "create table mt_merge(key UInt32) engine = Merge(default, '^mt_*')"
```

```
clickhouse benchmark <<< "select count() from mt_merge where toFloat64(key) > 500" --cumulative
```

This PR:
```
localhost:9000, queries 36, QPS: 6.174, RPS: 6174059.662, MiB/s: 47.104, result RPS: 6.174, result MiB/s: 0.000.

0.000%          0.152 sec.
10.000%         0.156 sec.
20.000%         0.156 sec.
30.000%         0.158 sec.
40.000%         0.159 sec.
50.000%         0.161 sec.
60.000%         0.162 sec.
70.000%         0.163 sec.
80.000%         0.164 sec.
90.000%         0.170 sec.
95.000%         0.175 sec.
99.000%         0.187 sec.
99.900%         0.187 sec.
99.990%         0.187 sec.
```

Current master:
```
localhost:9000, queries 9, QPS: 0.961, RPS: 960984.088, MiB/s: 7.332, result RPS: 0.961, result MiB/s: 0.000.

0.000%          1.013 sec.
10.000%         1.017 sec.
20.000%         1.020 sec.
30.000%         1.020 sec.
40.000%         1.037 sec.
50.000%         1.038 sec.
60.000%         1.044 sec.
70.000%         1.063 sec.
80.000%         1.063 sec.
90.000%         1.066 sec.
95.000%         1.067 sec.
99.000%         1.067 sec.
99.900%         1.067 sec.
99.990%         1.067 sec.
```