@CurtizJ 2020-11-13
I don't know how to create a perf test, because the query lasts more than 0.1s (recommended time) only when 1000+ tables are created, but creation of 1000 tables lasts more than 1 minute, which is unacceptable for perf tests.