@akuzm 2020-11-13
> I don't know how to create a perf test, because the query lasts more than 0.1s (recommended time) only when 1000+ tables are created, but creation of 1000 tables lasts more than 1 minute, which is unacceptable for perf tests.

I think in this case we can use a query that runs faster than 0.1s, because there is no practical way to increase its run time. We should mark it as "short" to avoid warnings. Not sure we will be able to detect the difference, but we can try.